backend_python/requirements.txt
fastapi
uvicorn
pillow
python-multipart
opencv-python
numpy
